third_party/blink/web_tests/external/wpt/css/css-gaps/tentative/parsing/gap-decorations-bidirectional-shorthands.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Gap Decoration: shorthand rule-* computed style values from longhands</title>
<link rel="help" href="https://tabatkins.github.io/specs/css-masonry/">
<link rel="author" title="Sam Davis Omekara Jr." href="mailto:samomekarajr@microsoft.com">
<script src="/resources/testharness.js" type="text/javascript"></script>
<script src="/resources/testharnessreport.js" type="text/javascript"></script>
</head>
<body>
<div id="target1"></div>
<div id="target2"></div>
<style>
  #target1 {
    column-rule-color: lime;
    row-rule-color: lime;
  }

  #target2 {
    column-rule-color: blue;
    row-rule-color: red;
  }
</style>
<script>
  test(function() {
    assert_equals(window.getComputedStyle(document.getElementById('target1')).getPropertyValue('rule-color'), 'rgb(0, 255, 0)');

  }, "rule-color shorthand computed from longhand values");

  test(function() {
    assert_equals(window.getComputedStyle(document.getElementById('target2')).getPropertyValue('rule-color'), '');
  }, "rule-color shorthand cannot be computed from longhand values so expect an empty string");
</script>
</body>
</html>
